+++ app/src/main/java/com/mark/pocketmanager/CustomClass/CategoryAmount.java
@@ -2,5 +2,5 @@ package com.mark.pocketmanager.CustomClass;
 
 public class CategoryAmount {
     public String Category;
-    public int Amount;
+    public long Amount;
 }
